refactor(ios): move user interface listener to another extension
